bundle: explicitely set workingDir instead of usind chdir()
this also fixes a bug with cmd_init in which repos-config was
created for the wrong workingDir (and e.g. with an not updated git repo
state)

this also fixes a bug with cmd_clone in which not all configuration
files were created - leading to a dirty git after './bundle apply' apply
on the server.