Consider highlighted suites to be supplier suites
Add highlighted suites to the suite-collection, which is subsequently
used to iterate over all suites, which shall be updated/queried/inspected.
The effect is that if a suite is defined by the user in the highlithted-set
only (by specifying --highlighted-suites) this suite will be considered to
be a supplier suite even if it was not explicitly defined as such
(by specifying --supplier-suites). Reference-suites are treated
differently, which is the implicit reason, why the highglighted suites
will be treated as supplier suites.